Post now has comments pulled in
Currently in it's most basic form. Not a fan of what I had to do to get it to work. a lot of function-based creation of content, but that's how it's gotta be for now 'til a new way is found.

Minor changes on:
-formatting
-db_dealer
-set

and new files
comment_loader
commentObject
and
get_comments
for instance.
The other stuff are just unused right now so cleanup is in order some time.